requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0
deep-translator>=1.11.0
geopy>=2.3.0
scikit-learn>=1.3.0
streamlit_folium>=0.13.0
polyline>=1.4.0
streamlit-drawable-canvas>=0.9.0
opencv-python-headless>=4.8.0
pandas>=2.0.0

# YOLO関連ライブラリ（必要に応じてコメントアウト）
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0